+++ packages/frontend/src/components/MkUrlPreview.vue
@@ -237,6 +237,7 @@ onMounted(async () => {
 		}
 	}
 });
+
 function truncateString(str) {
 	if (str.length > 100) {
 		return str.substring(0, 100) + '...';
