+++ packages/frontend/src/components/global/MkCustomEmoji.vue
@@ -36,7 +36,7 @@ import { customEmojis, customEmojisMap } from '@/custom-emojis.js';
 import * as os from '@/os.js';
 import { misskeyApiGet } from '@/scripts/misskey-api.js';
 import * as sound from '@/scripts/sound.js';
-import {copyToClipboard} from '@/scripts/copy-to-clipboard.js';
+import { copyToClipboard } from '@/scripts/copy-to-clipboard.js';
 import { i18n } from '@/i18n.js';
 import MkCustomEmojiDetailedDialog from '@/components/MkCustomEmojiDetailedDialog.vue';
 import { $i } from '@/account.js';
@@ -96,7 +96,7 @@ const alt = computed(() => `:${customEmojiName.value}:`);
 const errored = ref(url.value == null);
 
 function onClick(ev: MouseEvent) {
-	if (props.menu  && canReact.value ) {
+	if (props.menu && canReact.value ) {
 		os.popupMenu([{
 			type: 'label',
 			text: `:${props.name}:`,
