+++ packages/frontend/src/custom-emojis.ts
@@ -75,6 +75,7 @@ function toKanaToHira(str: string): string {
 		return String.fromCharCode(chr);
 	});
 }
+
 let cachedTags;
 export function getCustomEmojiTags() {
 	if (cachedTags) return cachedTags;
