+++ packages/frontend/src/pages/admin/custom-emojis-manager.local.list.vue
@@ -49,7 +49,6 @@ SPDX-License-Identifier: AGPL-3.0-only
 </template>
 
 <script lang="ts">
-import type { SortOrder } from '@/components/MkSortOrderEditor.define.js';
 import type { GridSortOrderKey } from './custom-emojis-manager.impl.js';
 
 export type EmojiSearchQuery = {
