+++ packages/frontend/src/pages/explore.vue
@@ -29,7 +29,7 @@ import MkFoldableSection from '@/components/MkFoldableSection.vue';
 import MkHorizontalSwipe from '@/components/MkHorizontalSwipe.vue';
 import { definePageMetadata } from '@/scripts/page-metadata.js';
 import { i18n } from '@/i18n.js';
-import {miLocalStorage} from "@/local-storage.js";
+import { miLocalStorage } from "@/local-storage.js";
 
 const props = withDefaults(defineProps<{
 	tag?: string;
